M: fixed some issues from the feedback
After the final boss was defeated, the character started snoring again.
The hurt sound was also activated again. These bugs have been fixed.

Display of the status bar of character and final boss has been adjusted.
Character now dies as soon as the life bar is at 0.
The life bar of the final boss no longer remains full after the first hit.

The touchbuttons are now available for the tablets.